letsreadarticles/requirements.txt
appdirs==1.4.4
astroid==2.4.2
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cs50==5.1.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.6
passlib==1.7.4
pylint==2.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.27
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
wrapt==1.12.1
